+++ MoMitClient/ip.txt
@@ -1,7 +1,2 @@
-It is a example file.
-you need change it to your own ip data.
-Min:2 servers High:10 servers
-It name is ip.txt.
-ip data like this:
-ip port publickey
-ip port publickey
+127.0.0.1 2448 fefe
+127.0.0.1 10876 fefe
