+++ MoMitClient/random.go
@@ -1,0 +1,7 @@
+package main
+
+func IV1(n int) {
+}
+
+func IV2(n int) {
+}
