+++ MoMitServer/104.28.160.198.crt
@@ -1,0 +1,10 @@
+-----BEGIN CERTIFICATE-----
+MIIBaTCCAQ+gAwIBAgIQO7r2FRmy5HPzVjhjdffPkTAKBggqhkjOPQQDAjAaMRgw
+FgYDVQQKEw9NeSBPcmdhbml6YXRpb24wIBcNMjUwMTA2MTE1MjU5WhgPMjEyNDEy
+MTMxMTUyNTlaMBoxGDAWBgNVBAoTD015IE9yZ2FuaXphdGlvbjBZMBMGByqGSM49
+AgEGCCqGSM49AwEHA0IABNYkKXBcmkiuvbjI/CWoTyCOPs4eaBTqQSqztlThT2F1
+edG5f489uTHpUBw6e7KMcXKvqpdQRl66vZpUQA2UMCujNTAzMA4GA1UdDwEB/wQE
+AwIFoDATBgNVHSUEDDAKBggrBgEFBQcDATAMBgNVHRMBAf8EAjAAMAoGCCqGSM49
+BAMCA0gAMEUCIACxmKzBxCA9fVrmA9tCJS6oq1nnoNPCSqEcGQ357CIdAiEA0tt0
+gOBTdtjbBkE0tLp4G1wPfgGFUh236Phy+671qSw=
+-----END CERTIFICATE-----
