+++ MoMitServer/MoMitServer.go
@@ -2,52 +2,45 @@ package main
 
 import (
 	"fmt"
-	"io/ioutil"
-	"net/http"
 	"os"
 )
 
-func getIP(url string) (string, error) {
-	resp, err := http.Get(url)
-	if err != nil {
-		return "", err
-	}
-	defer resp.Body.Close()
-
-	body, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return "", err
-	}
-
-	return string(body), nil
-}
 func main() {
-	ipv4, err := getIP("https://ipinfo.io/ip")
+	var ipv4, ipv6 string
+	var nocreatecert bool = false
+	var noCheckIP bool = false
+	config, err := readConfig("config.json")
 	if err != nil {
-		fmt.Println("Failed to get public IPv4: %w", err)
-		ipv4 = "Not available (IPv4 unsupported)"
-	}
-
-	ipv6, err := getIP("https://v6.ipinfo.io/ip")
-	if err != nil {
-		fmt.Println("Failed to get public IPv6: %w", err)
-		ipv6 = "Not available (IPv6 unsupported)"
-	}
-	fmt.Println(" __  __       __  __ _ _   ____")
-	fmt.Println("|  \\/  | ___ |  \\/  (_) |_/ ___| ")
-	fmt.Println("| |\\/| |/ _ \\| |\\/| | | __\\___ \\ / _ \\ '__\\ \\ / / _ \\ '__|")
-	fmt.Println("| |  | | (_) | |  | | | |_ ___) |  __/ |   \\ V /  __/ |")
-	fmt.Println("|_|  |_|\\___/|_|  |_|_|\\__|____/ \\___|_|    \\_/ \\___|_|")
-	fmt.Println("Server Public IPv4 Address:", ipv4)
-	fmt.Println("Server Public IPv6 Address:", ipv6)
-	if ipv4 == "Not available (IPv4 unsupported)" {
-		fmt.Println("Public IPv4 address is not available.")
-		if ipv6 == "Not available (IPv6 unsupported)" {
-			fmt.Println("No public IP address is available. Exiting.")
+		if os.IsNotExist(err) {
+			fmt.Println("INFO: config.json not found, default behavior will be applied.")
+		} else {
+			fmt.Println("Error reading config file:", err)
 			os.Exit(1)
 		}
+	} else {
+		if config.NocreateCert {
+			nocreatecert = true
+		}
 	}
-	generateCertificate(ipv4)
-	fmt.Println("Created a TLS certificate for Websocket connections by IPV4.")
-	fmt.Println("Waiting for incoming connections...")
+
+	if !noCheckIP {
+		ipv4, err = getIP("https://ipinfo.io/ip")
+		if err != nil {
+			fmt.Println("Failed to get public IPv4:", err)
+			ipv4 = "Not available"
+		}
+
+		ipv6, err = getIP("https://v6.ipinfo.io/ip")
+		if err != nil {
+			fmt.Println("Failed to get public IPv6:", err)
+			ipv6 = "Not available"
+		}
+	} else {
+		fmt.Println("INFO: IP address checking is disabled, so the certificate will not be generated.")
+		ipv4 = "Skipped"
+		ipv6 = "Skipped"
+		nocreatecert = true
+	}
+	rip := ipv4
+	printOutput(ipv4, ipv6, nocreatecert, rip)
 }
