+++ MoMitServer/ip.go
@@ -1,0 +1,21 @@
+package main
+
+import (
+	"io/ioutil"
+	"net/http"
+)
+
+func getIP(url string) (string, error) {
+	resp, err := http.Get(url)
+	if err != nil {
+		return "", err
+	}
+	defer resp.Body.Close()
+
+	body, err := ioutil.ReadAll(resp.Body)
+	if err != nil {
+		return "", err
+	}
+
+	return string(body), nil
+}
